--- source/app/about-us/about-us.component.html
@@ -1,1 +1,0 @@
-<p>about-us works!</p>
