+++ source/app/app.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
-import { HomeComponent } from './home/home.component';
-import { AboutUsComponent } from './about-us/about-us.component';
+import { HomeComponent } from './views/home/home.component';
+import { AboutUsComponent } from './views/about-us/about-us.component';
 
 export const routes: Routes = [
     { path: '', component: HomeComponent },
