--- source/app/home/home.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-import { FooterComponent } from '../shared/components/footer/footer.component';
-import { NavbarComponent } from '../shared/components/navbar/navbar.component';
-
-@Component({
-    selector: 'app-home',
-    standalone: true,
-    imports: [NavbarComponent, FooterComponent],
-    templateUrl: './home.component.html',
-    styleUrl: './home.component.css',
-})
-export class HomeComponent {}
