+++ source/app/views/about-us/about-us.component.html
@@ -1,0 +1,1 @@
+<p>about-us works!</p>
