+++ source/app/views/about-us/about-us.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-about-us',
+  standalone: true,
+  imports: [],
+  templateUrl: './about-us.component.html',
+  styleUrl: './about-us.component.css'
+})
+export class AboutUsComponent {
+
+}
